Moved onDownloadQuestion into QuestionBuilder
